+++ src/states/controlState.ts
@@ -60,7 +60,7 @@ export interface InfoCardElementStates {
   [key: string]: InfoCardElementState;
 }
 
-export type ItemFlyMode = "none" | "flyTo" | "flyToZoom";
+export type LocateMode = "doNothing" | "pan" | "panAndZoom";
 
 export type UserPlacesFormatName = "geojson" | "csv" | "wkt";
 
@@ -118,8 +118,8 @@ export interface ControlState {
   showDatasetBoundaries: boolean;
   baseMapUrl: string;
   showRgbLayer: boolean;
-  datasetFlyMode: ItemFlyMode;
-  placeFlyMode: ItemFlyMode;
+  datasetLocateMode: LocateMode;
+  placeLocateMode: LocateMode;
   exportTimeSeries: boolean;
   exportPlaces: boolean;
   exportTimeSeriesSeparator: string;
@@ -162,8 +162,8 @@ export function newControlState(): ControlState {
     mapInteraction: "Point",
     lastMapInteraction: "Point",
     showRgbLayer: false,
-    datasetFlyMode: "flyTo",
-    placeFlyMode: "flyToZoom",
+    datasetLocateMode: "pan",
+    placeLocateMode: "panAndZoom",
     volumeCardOpen: false,
     volumeRenderMode: "mip",
     volumeStates: {},